added resiliency to endpoints and clean up
