+++ backend/server.ts
@@ -13,10 +13,10 @@ import { currentUserRouter } from './routes/current-user';
 import { signinRouter } from './routes/signin';
 import { signoutRouter } from './routes/signout';
 import { signupRouter } from './routes/signup';
-import { postRouter } from './routes/post';
+import { postRouter } from './routes/post/post';
 import { errorHandler } from './middlewares/error-handler';
 import { NotFoundError } from './errors/not-found-error';
-import { deleteRouter } from './routes/delete';
+import { deleteRouter } from './routes/post/delete';
 
 // Configure env vars in env file
 require('dotenv').config();
